requirements.txt
torch==2.1.1
einops==0.7.0
pytest==7.4.3
pyhamcrest==2.1.0
numpy==1.26.2
